test(smoketest): Extract common shared Playwright context options
